[refac] move class to appropriate package
